pyproject.toml
[project]
name = "jpd-bot"
version = "0.1.0"
description = "Claude powered, RAG enhanced, policy review bot"
authors = [
    {name = "Martin Gleason",email = "marty.data.scientist@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "anthropic (>=0.76.0,<0.77.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pypdf (>=6.6.0,<7.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
